+++ FA2sp/Ext/CMapData/Body.h
@@ -9,6 +9,27 @@ struct OverlayTypeData
     bool Rock;
 };
 
+struct BuildingDataExt
+{
+    ~BuildingDataExt()
+    {
+        if (Foundations)
+            delete Foundations;
+        if (LinesToDraw)
+            delete LinesToDraw;
+    }
+
+    bool IsCustomFoundation() const
+    {
+        return Foundations != nullptr;
+    }
+
+    int Width{ 0 };
+    int Height{ 0 };
+    std::vector<MapCoord>* Foundations{ nullptr };
+    std::vector<std::pair<MapCoord, MapCoord>>* LinesToDraw{ nullptr };
+};
+
 class CMapDataExt : public CMapData
 {
 public:
@@ -25,6 +46,9 @@ public:
     int GetOreValueAt(CellData& cell);
     void InitOreValue();
 
+    void InitializeBuildingTypesExt(const char* ID);
+
+    static std::vector<BuildingDataExt> BuildingDataExts;
     static std::vector<OverlayTypeData> OverlayTypeDatas;
     static int OreValue[4];
 };